refactor: remove lib folder and rename Tailwind util
